우선 아주 취약한 인증 방법 수행
당장 Spring Security 를 도입하면 많이 번거로워지기 때문에,
아주 간단한 취약한 인증을 도입했음
--> Authorization 헤더로 사용자를 식별하는 방법

+++ src/main/java/com/example/vuelog/controller/PostController.java
@@ -4,7 +4,6 @@ import com.example.vuelog.dto.request.PostCreate;
 import com.example.vuelog.dto.request.PostEdit;
 import com.example.vuelog.dto.request.PostSearch;
 import com.example.vuelog.dto.response.PostResponse;
-import com.example.vuelog.exception.InvalidRequestException;
 import com.example.vuelog.service.PostService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -23,12 +22,12 @@ public class PostController {
     private final PostService postService;
 
     @PostMapping(path = "/posts")
-    public ResponseEntity<Void> post(@RequestBody @Valid PostCreate postCreate) {
-        postCreate.validate();
-        postService.write(postCreate);
-        return ResponseEntity
-                .status(HttpStatus.CREATED)
-                .build();
+    public void post(@RequestBody @Valid PostCreate postCreate,
+                     @RequestHeader String authorization) {
+        if (authorization.equals("revi1337")) {
+            postCreate.validate();
+            postService.write(postCreate);
+        }
     }
 
     @GetMapping(path = "/posts/{postId}")
@@ -46,21 +45,20 @@ public class PostController {
     }
 
     @PatchMapping(path = "/posts/{postId}")
-    public ResponseEntity<Void> updatePost(
-            @PathVariable(name = "postId") Long id, @RequestBody @Valid PostEdit postEdit) {
-        postService.edit(id, postEdit);
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .build();
+    public void updatePost(@PathVariable(name = "postId") Long id,
+                                           @RequestBody @Valid PostEdit postEdit,
+                                           @RequestHeader String authorization) {
+        if (authorization.equals("revi1337")) {
+            postService.edit(id, postEdit);
+        }
     }
 
     @DeleteMapping(path = "/posts/{postId}")
-    public ResponseEntity<Void> deletePost(@PathVariable(name = "postId") Long id) {
-        postService.delete(id);
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .build();
+    public void deletePost(@PathVariable(name = "postId") Long id,
+                                           @RequestHeader String authorization) {
+        if (authorization.equals("revi1337")) {
+            postService.delete(id);
+        }
     }
-
 
 }
